Added imports to Comment class fixed bug in CommentRepo with wrong methods name (finBY => findBy) Changed constructopr in Login information ( deleted id from constructor), added som imports Changed findByUserNameAAndPassword method (return value) Created MassageHandeleService and added som method to it for printing out massages with in color. Created login and loginOut methods. Changed start menu. Added login information to person class. Created createUser method.(Needs to refactor and divide into small parts) added som imports to UserTypeRepository
Fixed wrong inputs overall
