+++ src/main/java/com/example/wikipediagang/Model/LoginInformation.java
@@ -1,6 +1,6 @@
-package com.example.wikipediagang.model;
+package com.example.wikipediagang.Model;
 
-import com.example.wikipediagang.Person;
+import com.example.wikipediagang.Model.Person;
 import jakarta.persistence.*;
 
 @Entity
