+++ src/main/java/com/example/wikipediagang/model/ArticleHardCopy.java
@@ -1,4 +1,4 @@
-package com.example.wikipediagang.model;
+package com.example.wikipediagang.Model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,6 +7,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import  com.example.wikipediagang.Model.Article;
 
 @Entity
 @Table(name="article_hard_copy")
