+++ src/main/java/com/example/wikipediagang/repo/ArticleHardCopyRepo.java
@@ -1,6 +1,6 @@
 package com.example.wikipediagang.repo;
 
-import com.example.wikipediagang.model.ArticleHardCopy;
+import com.example.wikipediagang.Model.ArticleHardCopy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
