+++ src/main/java/com/example/wikipediagang/repo/ArticleRepo.java
@@ -1,7 +1,7 @@
 package com.example.wikipediagang.repo;
 
-import com.example.wikipediagang.model.Person;
-import com.example.wikipediagang.model.Article;
+import com.example.wikipediagang.Model.Person;
+import com.example.wikipediagang.Model.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
