+++ src/main/java/com/example/wikipediagang/repo/LoginInformationRepo.java
@@ -1,11 +1,8 @@
-package com.example.wikipediagang.Model;
+package com.example.wikipediagang.repo;
 
 import com.example.wikipediagang.Model.LoginInformation;
-import com.example.wikipediagang.Person;
-import com.example.wikipediagang.model.Person;
-import com.example.wikipediagang.model.LoginInformation;
+import com.example.wikipediagang.Model.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
