+++ src/main/java/com/example/wikipediagang/repo/PersonRepository.java
@@ -1,6 +1,6 @@
 package com.example.wikipediagang.repo;
 
-import com.example.wikipediagang.model.Person;
+import com.example.wikipediagang.Model.Person;
 import com.example.wikipediagang.model.UserType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
