+++ src/main/java/com/example/wikipediagang/repo/UserTypeRepository.java
@@ -4,10 +4,12 @@ import com.example.wikipediagang.model.UserType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 
 public interface UserTypeRepository  extends JpaRepository<UserType, Integer > {
 
-UserType findByName(String name);
+    Optional<UserType> findByName(String name);
 
 }
